supply_chain_env/server/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
openenv-core>=0.2.1
websockets>=12.0
python-multipart
